♻️ refactor: Moving the createTotpProvider file to the repositories folder.
